\name{checkExtension}
\alias{checkExtension}
\title{
checkExtension: checks for correct extensions of file names
}
\description{
This is an internal function of the GenePattern package.
}
\usage{
checkExtension(file.name, extension)
}
\arguments{
  \item{file.name }{
character vector specifying the file name to be written
}
  \item{extension }{
character vector specifying the extension to be checked
}
}
\value{
Returns a character string with the correct extension properly appended to the file name.
}
\references{
About the GenePattern package:
\url{
http://www.broadinstitute.org/cancer/software/genepattern/gp_guides/programmers/sections/gp_r
}
}
\author{
Gil Tomas
\email{gil.tomas@ulb.ac.be}
}