\name{MicroarrayToolbox}
\alias{MicroarrayToolbox-package}
\docType{package}
\title{MicroarrayToolbox: functions for routine microarray analysis}
\description{
These functions provide basic functionality to explore gene expression profiles data. In order to streamline and standardize analyses, microarray data is expected to be provided in the 'gct' format, proposed by the Broad Institute and described here:
http://www.broadinstitute.org/cancer/software/genepattern/gp_guides/file-formats/sections/gct
}
\details{
\tabular{ll}{
Package: \tab MicroarrayToolbox\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-08-07\cr
License: \tab GPL (>= 3)\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Gil Tomas and others.

Maintainer: Gil Tomas <gil.tomas@ulb.ac.be>
}