\name{checkExtension}
\alias{checkExtension}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
checkExtension: checks for correct extension of file names
}
\description{
This is an internal function of the GenePattern package.
}
\usage{
checkExtension(file.name, extension)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{file.name}{
character vector specifying the file name to be written
}
  \item{extension}{
character vector specifying the extension to be checked
}
}
\value{
Returns a character string with the correct extension properly appended to the file name.
}
\references{
See more details here:
\url{
http://www.broadinstitute.org/cancer/software/genepattern/gp_guides/programmers/sections/gp_r
}
}
\author{
Broad Institute
}